+++ spec/resources/numismatics/monograms/numismatics/monogram_decorator_spec.rb
@@ -4,7 +4,7 @@ require "rails_helper"
 RSpec.describe Numismatics::MonogramDecorator do
   subject(:decorator) { described_class.new(monogram) }
   let(:monogram) { FactoryBot.create_for_repository(:numismatic_monogram, member_ids: [file_set.id]) }
-  let(:file_set) { FactoryBot.create_for_repository(:file_set) }
+  let(:file_set) { FactoryBot.create_for_repository(:file_set, title: ["File Set 5"]) }
 
   describe "#attachable_objects" do
     it "does not allow attaching objects" do
@@ -35,6 +35,12 @@ RSpec.describe Numismatics::MonogramDecorator do
   describe "#title" do
     it "renders title as single value" do
       expect(decorator.title).to eq("Test Monogram")
+    end
+  end
+
+  describe "#decorated_filename" do
+    it "adds the filename" do
+      expect(decorator.decorated_filename).to eq("File Set 5")
     end
   end